+++ lib/roqua/support.rb
@@ -1,6 +1,8 @@
 require 'logger'
-require 'roqua/support/logging'
+require 'roqua/support/instrumentation'
+require 'roqua/support/log_wrapper'
 require 'roqua/support/errors'
+require 'roqua/support/stats'
 
 module Roqua
   class << self
@@ -18,6 +20,14 @@ module Roqua
 
     def logger=(logger)
       @logger = LogWrapper.new(logger)
+    end
+
+    def stats
+      @stats ||= Stats.new
+    end
+
+    def stats=(stats)
+      @stats = stats
     end
   end
 end